+++ app/src/main/AndroidManifest.xml
@@ -37,7 +37,7 @@
         </activity>
 
         <activity
-            android:name=".ui.mypage.MypageChangePwActivity"
+            android:name="com.example.fit_i.ui.main.mypage.MypageChangePwActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -111,7 +111,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
