+++ app/src/main/java/com/example/fit_i/ui/main/MainActivity.kt
@@ -1,14 +1,13 @@
-package com.example.fit_i
+package com.example.fit_i.ui.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.os.PersistableBundle
-import androidx.fragment.app.FragmentManager
-import androidx.fragment.app.FragmentTransaction
+import com.example.fit_i.CommunityFragment
+import com.example.fit_i.R
 import com.example.fit_i.databinding.ActivityMainBinding
-import com.example.fit_i.ui.chat.ChatFragment
-import com.example.fit_i.ui.home.HomeFragment
-import com.example.fit_i.ui.mypage.MypageFragment
+import com.example.fit_i.ui.main.chat.ChatFragment
+import com.example.fit_i.ui.main.home.HomeFragment
+import com.example.fit_i.ui.main.mypage.MypageFragment
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
