+++ app/src/main/java/com/example/fit_i/ui/main/home/HomeFriendFragment.kt
@@ -1,4 +1,4 @@
-package com.example.fit_i.ui.home
+package com.example.fit_i.ui.main.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [HomeMedicalFragment.newInstance] factory method to
+ * Use the [HomeFriendFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class HomeMedicalFragment : Fragment() {
+class HomeFriendFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class HomeMedicalFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home_medical, container, false)
+        return inflater.inflate(R.layout.fragment_home_friend, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class HomeMedicalFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment HomeMedicalFragment.
+         * @return A new instance of fragment HomeFriendFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            HomeMedicalFragment().apply {
+            HomeFriendFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
