+++ app/src/main/java/com/example/fit_i/ui/main/mypage/MypageNoticeIngeFragment.kt
@@ -1,24 +1,26 @@
-package com.example.fit_i.ui.mypage
+package com.example.fit_i.ui.main.mypage
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageButton
-import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentTransaction
 import com.example.fit_i.R
 
-class MypagepermissonFragment : Fragment(){
-    private lateinit var binding: MypagepermissonFragment
+class MypageNoticeIngeFragment : Fragment() {
+    private lateinit var binding: MypageNoticeIngeFragment
+
 
     override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
+        inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val view = inflater.inflate(R.layout.fragment_mypagepermisson,container,false)
-        val ibpre = view.findViewById<View>(R.id.ib_pre3) as ImageButton
+
+        val ee = inflater.inflate(R.layout.fragment_mypage_notice_ing, container, false)
+        val ibpre = ee.findViewById<View>(R.id.ib_pre7) as ImageButton
+
 
         ibpre.setOnClickListener{
             val mypageFragment = MypageFragment()
@@ -30,7 +32,8 @@ class MypagepermissonFragment : Fragment(){
 
         }
 
-        return view
+        return ee
+
     }
 
 }
