+++ app/src/main/java/com/example/fit_i/ui/main/mypage/MypageSettingFragment.kt
@@ -1,23 +1,15 @@
-package com.example.fit_i.ui.mypage
+package com.example.fit_i.ui.main.mypage
 
 import android.annotation.SuppressLint
-import android.app.Activity
-import android.app.Dialog
 import android.os.Bundle
-import android.os.ProxyFileDescriptorCallback
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.ImageButton
 import androidx.appcompat.widget.AppCompatButton
-import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
 import com.example.fit_i.ProfileDialog
 import com.example.fit_i.R
-import com.example.fit_i.ui.mypage.MypageFragment
-import com.example.fit_i.ui.mypage.MypageSiginoutFragment
 
 class MypageSettingFragment :Fragment() {
     private lateinit var binding: MypageSettingFragment
