+++ app/src/main/java/com/example/fit_i/ui/main/mypage/MypagepermissonFragment.kt
@@ -1,26 +1,24 @@
-package com.example.fit_i.ui.mypage
+package com.example.fit_i.ui.main.mypage
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageButton
+import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentTransaction
 import com.example.fit_i.R
 
-class MypageNoticeIngeFragment : Fragment() {
-    private lateinit var binding: MypageNoticeIngeFragment
-
+class MypagepermissonFragment : Fragment(){
+    private lateinit var binding: MypagepermissonFragment
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-
-        val ee = inflater.inflate(R.layout.fragment_mypage_notice_ing, container, false)
-        val ibpre = ee.findViewById<View>(R.id.ib_pre7) as ImageButton
-
+        val view = inflater.inflate(R.layout.fragment_mypagepermisson,container,false)
+        val ibpre = view.findViewById<View>(R.id.ib_pre3) as ImageButton
 
         ibpre.setOnClickListener{
             val mypageFragment = MypageFragment()
@@ -32,8 +30,7 @@ class MypageNoticeIngeFragment : Fragment() {
 
         }
 
-        return ee
-
+        return view
     }
 
 }
