+++ app/src/main/res/layout/activity_change_pw.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.mypage.MypageChangePwActivity">
+    tools:context=".ui.main.mypage.MypageChangePwActivity">
 
     <ImageButton
         android:id="@+id/iv_back"
