+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.HomeFragment">
+    tools:context=".ui.main.home.HomeFragment">
 <!--    <ImageView-->
 <!--        android:id="@+id/iv_mainComment"-->
 <!--        android:layout_width="wrap_content"-->
