+++ app/src/main/res/layout/fragment_home_diet.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.HomeDietFragment">
+    tools:context=".ui.main.home.HomeDietFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
