+++ app/src/main/res/layout/fragment_home_eating.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.HomeEatingFragment">
+    tools:context=".ui.main.home.HomeEatingFragment">
 
 
     <!-- TODO: Update blank fragment layout -->
