+++ app/src/main/res/layout/fragment_home_friend.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.HomeFriendFragment">
+    tools:context=".ui.main.home.HomeFriendFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
