+++ app/src/main/res/layout/fragment_home_medical.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.HomeMedicalFragment">
+    tools:context=".ui.main.home.HomeMedicalFragment">
 
 
     <!-- TODO: Update blank fragment layout -->
