+++ app/src/main/res/layout/fragment_home_pt.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.HomePtFragment">
+    tools:context=".ui.main.home.HomePtFragment">
 
     <LinearLayout
         android:id="@+id/ll_sort"
