+++ app/src/main/res/layout/fragment_mypage_modify_ing.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.mypage.MypageModifyIngFragment">
+    tools:context=".ui.main.mypage.MypageModifyIngFragment">
 
     <ImageButton
         android:id="@+id/ib_pre"
