+++ app/src/main/res/layout/fragment_mypage_notice.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.mypage.MypageNoticeFragment">
+    tools:context=".ui.main.mypage.MypageNoticeFragment">
