+++ app/src/main/res/layout/fragment_mypage_siginout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.mypage.MypageSettingFragment">
+    tools:context=".ui.main.mypage.MypageSettingFragment">
 
     <ImageView
         android:id="@+id/logout"
